Refactor: use a method to convert json to nodes
I made a recursive method that walks the json and creates a
nodes tree rather than using the BeliefNode constructor.

The _growTree method and a subclass of BeliefNode (_MutableBeliefNode)
are private so that the recursive function can mutate BeliefNodes during
construction but provide a safer, ie mostly immutable, public interface.

I also added a models folder so that we now have four groups for classes
- beliefs, cognitions, models, views - which is a useful categerisation
that will make code more findable, as the classes fit into the categories
fairly intuitively (at least imo).